refactor: split SwitchDatabase type
